type_dict now propagates to leaves in hierarchical mode
Whereas (de)serializer_name_dict should exactly match the names of 
parameters in a hierarchy to avoid ambiguity, it's unlikely that a user 
would want to specify a child when using (de)serializer_type_dict. The 
functionality has been changed to allow for key-value pairs of the type 
dict to propagate to children of the hierarchy when the key is a class 
(not a str), but will drill-down the hierarchy whenever the key is a 
string.